+++ rest/src/main/kotlin/builder/interaction/PublicInteractionBuilder.kt
@@ -4,8 +4,6 @@ import dev.kord.common.annotation.KordDsl
 import dev.kord.common.annotation.KordPreview
 import dev.kord.common.entity.AllowedMentions
 import dev.kord.common.entity.InteractionResponseType
-import dev.kord.common.entity.MessageFlag
-import dev.kord.common.entity.MessageFlags
 import dev.kord.common.entity.optional.Optional
 import dev.kord.common.entity.optional.OptionalBoolean
 import dev.kord.common.entity.optional.delegate.delegate
@@ -23,24 +21,6 @@ import java.nio.file.Path
 import kotlin.contracts.ExperimentalContracts
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
-
-
-@KordPreview
-@KordDsl
-class AcknowledgementResponseBuilder(val ephemeral: Boolean = false) :
-    RequestBuilder<InteractionResponseCreateRequest> {
-    override fun toRequest(): InteractionResponseCreateRequest {
-        val data = if (!ephemeral) Optional.Missing()
-        else Optional.Value(
-            InteractionApplicationCommandCallbackData(flags = Optional.Value(MessageFlags(MessageFlag.Ephemeral)))
-        )
-        return InteractionResponseCreateRequest(
-            InteractionResponseType.DeferredChannelMessageWithSource,
-            data
-        )
-    }
-
-}
 
 @KordPreview
 @KordDsl
@@ -85,9 +65,13 @@ class PublicInteractionResponseCreateBuilder :
     }
 
     override fun toRequest(): MultipartInteractionResponseCreateRequest {
+        val type =
+            if (files.isEmpty() && content == null && embeds == null) InteractionResponseType.DeferredChannelMessageWithSource
+            else InteractionResponseType.ChannelMessageWithSource
+
         return MultipartInteractionResponseCreateRequest(
             InteractionResponseCreateRequest(
-                InteractionResponseType.ChannelMessageWithSource,
+                type,
                 InteractionApplicationCommandCallbackData(
                     content = _content,
                     embeds = _embeds.mapList { it.toRequest() },
@@ -113,7 +97,6 @@ class PublicInteractionResponseModifyBuilder :
 
     private var _allowedMentions: Optional<AllowedMentions> = Optional.Missing()
     var allowedMentions: AllowedMentions? by ::_allowedMentions.delegate()
-
 
     val files: MutableList<Pair<String, InputStream>> = mutableListOf()
 
@@ -141,73 +124,18 @@ class PublicInteractionResponseModifyBuilder :
     }
 
     override fun toRequest(): MultipartInteractionResponseModifyRequest {
+        val type =
+            if (files.isEmpty() && content == null && embeds == null) InteractionResponseType.DeferredChannelMessageWithSource
+            else InteractionResponseType.ChannelMessageWithSource
+
         return MultipartInteractionResponseModifyRequest(
             InteractionResponseModifyRequest(
-                content = _content,
-                _embeds.mapList { it.toRequest() },
-                _allowedMentions
+                    content = _content,
+                    embeds = _embeds.mapList { it.toRequest() },
+                    allowedMentions = _allowedMentions,
             ),
             files
         )
+
     }
 }
-
-
-@KordPreview
-@KordDsl
-class EphemeralInteractionResponseModifyBuilder() :
-    RequestBuilder<MultipartInteractionResponseModifyRequest> {
-    private var _content: Optional<String> = Optional.Missing()
-    var content: String? by ::_content.delegate()
-    private var _allowedMentions: Optional<AllowedMentions> = Optional.Missing()
-    var allowedMentions: AllowedMentions? by ::_allowedMentions.delegate()
-
-    @OptIn(ExperimentalContracts::class)
-    inline fun allowedMentions(builder: AllowedMentionsBuilder.() -> Unit) {
-        contract { callsInPlace(builder, InvocationKind.EXACTLY_ONCE) }
-        allowedMentions = AllowedMentionsBuilder().apply(builder).build()
-    }
-
-
-    override fun toRequest(): MultipartInteractionResponseModifyRequest {
-        return MultipartInteractionResponseModifyRequest(
-            InteractionResponseModifyRequest(
-                content = _content,
-                allowedMentions = _allowedMentions,
-            ),
-        )
-    }
-}
-
-
-@KordPreview
-@KordDsl
-class EphemeralInteractionResponseCreateBuilder(val content: String) :
-    RequestBuilder<InteractionResponseCreateRequest> {
-    private var _allowedMentions: Optional<AllowedMentions> = Optional.Missing()
-    var allowedMentions: AllowedMentions? by ::_allowedMentions.delegate()
-
-    private var _tts: OptionalBoolean = OptionalBoolean.Missing
-    var tts: Boolean? by ::_tts.delegate()
-
-    @OptIn(ExperimentalContracts::class)
-    fun allowedMentions(builder: AllowedMentionsBuilder.() -> Unit) {
-        contract { callsInPlace(builder, InvocationKind.EXACTLY_ONCE) }
-        allowedMentions = AllowedMentionsBuilder().apply(builder).build()
-    }
-
-
-    override fun toRequest(): InteractionResponseCreateRequest {
-        return InteractionResponseCreateRequest(
-            type = InteractionResponseType.ChannelMessageWithSource,
-            InteractionApplicationCommandCallbackData(
-                content = Optional.Value(content),
-                allowedMentions = _allowedMentions,
-                tts = _tts,
-                flags = Optional.Value(MessageFlags(MessageFlag.Ephemeral))
-
-            ).optional()
-        )
-    }
-}
-
